requirements.txt
cffi==0.9.2
cryptography==0.8.1
enum34==1.0.4
ndg-httpsclient==0.3.3
peewee==2.5.0
pyasn1==0.1.7
pycparser==2.10
PyMySQL==0.6.6
pyOpenSSL==0.14
requests==2.6.0
six==1.9.0